+++ include/igraph_cocitation.h
@@ -43,16 +43,16 @@ IGRAPH_EXPORT igraph_error_t igraph_bibcoupling(const igraph_t *graph, igraph_ma
                                      const igraph_vs_t vids);
 
 IGRAPH_EXPORT igraph_error_t igraph_similarity_jaccard(const igraph_t *graph, igraph_matrix_t *res,
-                                            const igraph_vs_t vids, igraph_neimode_t mode,
-                                            igraph_bool_t loops);
+                                            const igraph_vs_t vit_from, const igraph_vs_t vit_to,
+                                            igraph_neimode_t mode, igraph_bool_t loops);
 IGRAPH_EXPORT igraph_error_t igraph_similarity_jaccard_pairs(const igraph_t *graph, igraph_vector_t *res,
                                                   const igraph_vector_int_t *pairs, igraph_neimode_t mode, igraph_bool_t loops);
 IGRAPH_EXPORT igraph_error_t igraph_similarity_jaccard_es(const igraph_t *graph, igraph_vector_t *res,
                                                const igraph_es_t es, igraph_neimode_t mode, igraph_bool_t loops);
 
 IGRAPH_EXPORT igraph_error_t igraph_similarity_dice(const igraph_t *graph, igraph_matrix_t *res,
-                                         const igraph_vs_t vids, igraph_neimode_t mode,
-                                         igraph_bool_t loops);
+                                         const igraph_vs_t vit_from, const igraph_vs_t vit_to,
+                                         igraph_neimode_t mode, igraph_bool_t loops);
 IGRAPH_EXPORT igraph_error_t igraph_similarity_dice_pairs(const igraph_t *graph, igraph_vector_t *res,
                                                const igraph_vector_int_t *pairs, igraph_neimode_t mode, igraph_bool_t loops);
 IGRAPH_EXPORT igraph_error_t igraph_similarity_dice_es(const igraph_t *graph, igraph_vector_t *res,