+++ src/misc/cocitation.c
@@ -299,9 +299,11 @@ static igraph_error_t igraph_i_neisets_intersect(
  * \param graph The graph object to analyze
  * \param res Pointer to a matrix, the result of the calculation will
  *        be stored here. The number of its rows and columns is the same
- *        as the number of vertex IDs in \p vids.
- * \param vids The vertex IDs of the vertices for which the
- *        calculation will be done.
+ *        as the number of vertex IDs in \p vit_from and \p vit_to, respectively.
+ * \param vit_from The vertex IDs of the first set of vertices of the
+ *        pairs for which the calculation will be done.
+ * \param vit_to The vertex IDs of the second set of vertices of the
+ *        pairs for which the calculation will be done.
  * \param mode The type of neighbors to be used for the calculation in
  *        directed graphs. Possible values:
  *        \clist
@@ -335,7 +337,7 @@ static igraph_error_t igraph_i_neisets_intersect(
  * \example examples/simple/igraph_similarity.c
  */
 igraph_error_t igraph_similarity_jaccard(const igraph_t *graph, igraph_matrix_t *res,
-                              const igraph_vs_t vids, igraph_neimode_t mode, igraph_bool_t loops) {
+                              const igraph_vs_t vit_from, const igraph_vs_t vit_to, igraph_neimode_t mode, igraph_bool_t loops) {
     igraph_lazy_adjlist_t al;
     igraph_vit_t vit, vit2;
     igraph_integer_t i, j;
@@ -343,15 +345,15 @@ igraph_error_t igraph_similarity_jaccard(const igraph_t *graph, igraph_matrix_t 
     igraph_vector_int_t *v1, *v2;
     igraph_integer_t k;
 
-    IGRAPH_CHECK(igraph_vit_create(graph, vids, &vit));
+    IGRAPH_CHECK(igraph_vit_create(graph, vit_from, &vit));
     IGRAPH_FINALLY(igraph_vit_destroy, &vit);
-    IGRAPH_CHECK(igraph_vit_create(graph, vids, &vit2));
+    IGRAPH_CHECK(igraph_vit_create(graph, vit_to, &vit2));
     IGRAPH_FINALLY(igraph_vit_destroy, &vit2);
 
     IGRAPH_CHECK(igraph_lazy_adjlist_init(graph, &al, mode, IGRAPH_NO_LOOPS, IGRAPH_NO_MULTIPLE));
     IGRAPH_FINALLY(igraph_lazy_adjlist_destroy, &al);
 
-    IGRAPH_CHECK(igraph_matrix_resize(res, IGRAPH_VIT_SIZE(vit), IGRAPH_VIT_SIZE(vit)));
+    IGRAPH_CHECK(igraph_matrix_resize(res, IGRAPH_VIT_SIZE(vit), IGRAPH_VIT_SIZE(vit2)));
 
     if (loops) {
         for (IGRAPH_VIT_RESET(vit); !IGRAPH_VIT_END(vit); IGRAPH_VIT_NEXT(vit)) {
@@ -591,9 +593,11 @@ igraph_error_t igraph_similarity_jaccard_es(const igraph_t *graph, igraph_vector
  * \param graph The graph object to analyze.
  * \param res Pointer to a matrix, the result of the calculation will
  *        be stored here. The number of its rows and columns is the same
- *        as the number of vertex IDs in \p vids.
- * \param vids The vertex IDs of the vertices for which the
- *        calculation will be done.
+ *        as the number of vertex IDs in \p vit_from and \p vit_to, respectively.
+ * \param vit_from The vertex IDs of the first vertices of the
+ *        pairs for which the calculation will be done.
+ * \param vit_to The vertex IDs of the second vertices of the
+ *        pairs for which the calculation will be done.
  * \param mode The type of neighbors to be used for the calculation in
  *        directed graphs. Possible values:
  *        \clist
@@ -627,10 +631,10 @@ igraph_error_t igraph_similarity_jaccard_es(const igraph_t *graph, igraph_vector
  * \example examples/simple/igraph_similarity.c
  */
 igraph_error_t igraph_similarity_dice(const igraph_t *graph, igraph_matrix_t *res,
-                                      const igraph_vs_t vids,
+                                      const igraph_vs_t vit_from, const igraph_vs_t vit_to,
                                       igraph_neimode_t mode, igraph_bool_t loops) {
 
-    IGRAPH_CHECK(igraph_similarity_jaccard(graph, res, vids, mode, loops));
+    IGRAPH_CHECK(igraph_similarity_jaccard(graph, res, vit_from, vit_to, mode, loops));
 
     igraph_integer_t nr = igraph_matrix_nrow(res);
     igraph_integer_t nc = igraph_matrix_ncol(res);